Exercicise 16 vector solved
Make a program that reads a vector of 5 positions for real numbers, and then a 
integer code. If the code is zero, end the program; if 1, show the vector in order 
Direct; If it is 2, show the vector in reverse order. If the code is different from 1 and 2 
Write a message stating that the code is invalid.