Exercicise 1 Vector solved
Make a program that has a vector named A that stores 6 numbers inte- ́
Ros. The program should perform the following steps:
(a) Assign the following values to this vector: 1, 0, 5, -2, -5, 7.
(b) Store in an integer (simple) variable the sum of the values of the positions
A[0], A[1] and A[5] of the vector and show this sum on the screen.
(c) modify the vector in heading 4 by assigning to this heading the value 100. ˜
(d) Show on the screen each value of vector A, one in each line.