Exercise 17 and 18 Vetor Solved
Exercise17 - Read a vector of 10 positions and assign a value of 0 to all elements that have
negative values.
Exercicise18 - Make a program that reads a vector of 10 numbers. Read a n'umero' x. Count the
Multiples of an integer n ́ integer ́ x in a vector and show them on the screen.